+++ CHANGES.rst
@@ -17,6 +17,7 @@ Major Changes
 Minor Changes
 ^^^^^^^^^^^^^
 
+- Improvements & fixes to package tests.
 - Improvements & fixes to package documentation.